Added support for Trailers from ShowTimes. Can be accessed from both details view and main movie list.
Added support for a secondary trailer button (11900), this is needed for better skin navigation when multiple views share share the same window e.g. MovingPictures and Showtimes use the same window for the main movie list and details view. This means skinners can use either the primary id '11899' or '11900' or both.